refactor generate run kwargs
